[IMP] product: split python files by classes
